+++ src/components/QuoteListItem.vue
@@ -1,15 +1,104 @@
 <template>
-  <div>My component</div>
+  <q-item>
+    <q-item-main>
+      <p>{{ quote.text }}</p>
+      <q-item-tile sublabel lines="3" v-if="quote.notes !== '' && showNotes">{{ quote.notes }}</q-item-tile>
+      <br/>
+      <span>
+        <q-icon name="fa-pencil" @click.native="openEdit" class="cursor-pointer" />&nbsp;
+      </span>
+      <span v-if="showLocation && quote.locationType !== 'None'" class="q-item-sublabel">
+        {{ quote.locationType }} {{ quote.location }}&nbsp;
+      </span>
+      <span v-if="showTags">
+        <q-chip v-for="tag in quote.tags" small color="primary" :key="tag">{{ tag }}</q-chip>
+      </span>
+      <span v-if="showTags && showBible && quote.tags.length !== 0 && quote.bibleRefs.length !== 0">
+         |
+      </span>
+      <span v-if="showBible">
+        <q-chip v-for="bibleRef in quote.bibleRefs" small color="secondary" :key="bibleRef._id">{{ refString(bibleRef) }}</q-chip>
+      </span>
+    </q-item-main>
+    <q-item-side right v-if="showCover" style="margin: 20px">
+      <q-item-tile v-if="showCover" image>
+        <img :src="quote.mediaImageURL" class="cover-image" />
+      </q-item-tile>
+      <q-item-tile sublabel lines="1" v-if="showAuthor">{{ quote.author }}</q-item-tile>
+    </q-item-side>
+    <q-modal v-model="editOpen" content-classes="edit-quote-modal">
+      <q-icon name="fa-close" size="2rem" @click="editOpen=false" class="float-right cursor-pointer" />
+      <quote-form ref="quoteForm" :quote="quote" :modal-fin="closeEdit" form-type="Edit" />
+    </q-modal>
+  </q-item>
 </template>
 
 <script>
+import QuoteForm from 'components/QuoteForm.vue'
+
 export default {
-  // name: 'ComponentName',
+  components: {
+    QuoteForm
+  },
+  props: ['quoteObj', 'cover', 'author', 'location', 'tags', 'bible', 'notes'],
   data () {
-    return {}
+    return {
+      showCover: (this.cover === ''),
+      showAuthor: (this.author === ''),
+      showLocation: (this.location === ''),
+      showTags: (this.tags === ''),
+      showBible: (this.bible === ''),
+      showNotes: (this.notes === ''),
+      editOpen: false,
+      quote: this.quoteObj
+    }
+  },
+  watch: {
+    quote (value) {
+      this.quote = value
+    }
+  },
+  methods: {
+    refString (ref) {
+      if (ref.type === 'single') {
+        return ref.book + ' ' + ref.chapter + ':' + ref.verse
+      } else if (ref.type === 'simpleRange') {
+        return ref.book + ' ' + ref.chapter + ':' + ref.verse + '-' + ref.verser
+      } else if (ref.type === 'complexRange') {
+        return ref.book + ' ' + ref.chapter + ':' + ref.verse + '-' + ref.chapterr + ':' + ref.verser
+      } else {
+        console.error('Something went wrong with a Bible Ref...')
+      }
+    },
+    openEdit () {
+      this.editOpen = true
+      this.$refs.quoteForm.init(false)
+    },
+    closeEdit () {
+      this.editOpen = false
+    }
   }
 }
 </script>
 
 <style>
+
+.q-item-image {
+  margin-left: 20px;
+}
+.q-item-image img {
+  max-width: 50%;
+  max-height: 50%;
+}
+.edit-quote-modal {
+  padding: 30px;
+  width: 100%;
+}
+@media screen and (min-width: 1200px) {
+  .edit-quote-modal {
+    min-width: 500px;
+    width: 500px;
+  }
+}
+
 </style>